Solve the problem of skweezed rows
The way I deleted the rows that correspond to the overlap was not good. Every time a row is deleted, the size of the image reduces. The row 1 becomes the row 0, row 2 becomes the row 1, etc. The problem is now fixed by deleting only the row 0 with an iteration corresponding to the overlap percentage.